+++ src/main/java/chess/database/DBConnection.java
@@ -5,7 +5,7 @@ import java.sql.DriverManager;
 
 public class DBConnection {
     private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "root";
+    private static final String USER = "user";
     private static final String PASSWORD = "password";
 
     public static Connection getConnection() {
