--- src/main/java/chess/database/factory/RoomFactory.java
@@ -1,13 +1,0 @@
-package chess.database.factory;
-
-import chess.dao.RoomDao;
-import chess.database.DBConnection;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public class RoomFactory {
-
-}
